Add automatic download resume support
This patch adds automatic download resume support.
Individual segment files are written to the output
directory using the following format:

    .{emission ID}-{episode ID}-{bitrate}-{seg index}.ts

When one of those files exists, it's considered a
complete segment file, and will be skipped by the
downloader if found. When the downloader knows all
the segment files exists, it concatenates them as a
single output file (using the desired output file
name), and removes all the temporary segment files.