Allow specifying a single URL to fetch, list, info
This patch makes it possible to fetch, list or show the info simply by
specifying the URL to the resource (show or episode).  Therefore, it
should now be possible to fetch using one of the following forms:

  - toutv fetch <show>
  - toutv fetch <show> <episode>
  - toutv fetch <show-url>
  - toutv fetch <episode> <url>

Same for info.

We suggested some time ago to try to migrate from "emission" to "show",
since it's a bit silly (although cute) to use franglais.  I am
converting the areas I touch to use that name.

I removed the current test, since it's not very useful, and added a test
for the show/episode argument parsing.  It will be stable, since it
doesn't depend on external factors.