Split Downloader, introduce SegmentHandler
The listing/searching part has a relatively nice abstraction layer
(Transport/JsonTransport) that allows us to mock it in order to test
some components individually.  Unfortunately, the segment download part
does not have a similar thing, everything is in the Downloader class.
It makes HTTP requests and saves the data on the filesystem directly.

This commit is a first attempt at "modularizing" it, by introducing the
concept of a SegmentHandler.  A SegmentHandler is responsible of
handling the downloaded segments.  For now, the only implementation is
FilesystemSegmentHandler, which keeps the same behavior we have
currently: save each segment in a temporary file and stitch them all
together at the end.

The plan is to do the same for the segment source, adding a
SegmentProducer or SegmentProvider or whatever.  This way, we'll be
able to mock both the input and output and test various scenarios.

To keep the feature of being able to interrupt and resume downloads
(that's why temporary files for segments are used), I went through a few
hoops and added the has_segment/segment_size methods to the
SegmentHandler.  I don't find it particularly nice, since I don't think
it's the SegmentHandler's job do to this.