Update Python version check to 3.3+
Since we now use pytest for testing, and it requires Python 3.3+, update
our version check to 3.3+.  Most users won't ever run the tests, so they
should be able to run pytoutv with Python 3.2, but it's simpler for us
to require 3.3 for everything.  Python 3.2 installs should not be that
common anymore.