Add tests against the real tou.tv service
This patch is a modified version of gboudreau's more-tests branch.  It
adds tests that use the real tou.tv service.  Because of the possibility
of network errors, geographically restricted content and content
changing with time, we should expect intermittent failures.  For this
reason, this new test is skipped by default, as it should not be run by
Travis.