Improve show querying/matching in the client.
This refactor was initiated by issue #95:

  $ toutv fetch
  http://ici.tou.tv/en-audition-avec-simon/S01E08\?lectureauto\=1
  Cannot find "en-audition-avec-simon"
  Did you mean one of the following?

    * EN AUDITION AVEC SIMON
    * EN AUDITION AVEC SIMON - BENOÎT CLICHE

The problem is that to download episodes from URL, we are now extracting
the info directly from the URL.  In this example, we extract these
values:

  - show: en-audition-avec-simon
  - episode: S01E08

When en-audition-avec-simon is passed to get_emission_by_name, no
corresponding shows are found.  The candidates corresponding to that
show are:

  - "EN AUDITION AVEC SIMON"
  - 1436412334

So first, this patch adds to the candidates the names of the shows that
we find in the URL, "en-audition-avec-simon" in this case.  This
information is already available in the Emission object, except we need
to strip a loading forward slash.

This change allows to have an exact match from the URL.

Finally, this patch refactors the function a bit to use a dict that
maps candidates to the show they come from, so that it's easier to
return the show when we have an exact match.

The next step would be to improve how we report when there is no exact
match.  For example, we would now get this message, which is not that
nice:

  Cannot find "en audition avec simo"
  Did you mean one of the following?

    * EN AUDITION AVEC SIMON
    * EN-AUDITION-AVEC-SIMON
    * EN AUDITION AVEC SIMON - BENOÎT CLICHE

We could instead pass the Emission objects corresponding to the close
matches to the NoMatchException, so that the error messages could be
more nicely formatted.

Oh and it renames get_emission_by_name to get_emission_by_whatever,
since the old name was not very accurate...

Fixes #95