Bugfix: Timeout error during search
Search is the only function that still use the old API. That old API sometimes simply times out.

Error is:
Timeout error (20 s for "https://api.tou.tv/v1/toutvapiservice.svc/json/SearchTerms?query=...)

To work around this, we should lower the timeout for that API call, and just retry it a few times if we get Timeout errors.

Fixes #98