Print date in episode list
I think it would be useful to have the date in the episode list.  It
helps to know whether the latest episode is new.  Also, I think we can
hide the Id.  It's much more user friendly to refer to episodes using
the SAE.

Before:

  * 2425172033 - S17E25 - Épisode 25

After:

  * S17E25 - Épisode 25 - 9 mars 2017

The date comes already formatted like this from the JSON (in french).
Ideally, we should parse and to make it a "date" Python object and print
it using the user locale, but I think this is good enough for now.