tests: Call setup.py test instead of pytest directly
With Python 3.3 on Travis, pytest is not found.  Going through setup.py
to run the tests should make it install pytest, because it is specified
in tests_require, in setup.py.

With Python nightly on Travis, I am getting this error:

pkg_resources.DistributionNotFound: The 'setuptools_scm>=1.15.0' distribution was not found and is required by the application

The workaround is to install it manually.