Handle shows with no seasons, single episode
Some shows (https://ici.tou.tv/arthur-laventurier-au-costa-rica, at the
time of writing) have a single episode, and their SeasonLineups
attribute is None.  We can still download them if we treat the top-level
object as the episode (the important value is idMedia, which we will
eventually use to get the playlist of video segments).  This patch
attempts to do this.